# generated/src/main/java/org/openapitools/model/MemoryDTO.java
package org.openapitools.model;

import java.net.URI;
import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonValue;
import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;
import org.openapitools.model.Image;
import org.openapitools.model.Mood;
import org.openapitools.jackson.nullable.JsonNullable;
import java.time.OffsetDateTime;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import io.swagger.v3.oas.annotations.media.Schema;


import java.util.*;
import jakarta.annotation.Generated;

/**
 * MemoryDTO
 */

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-11T18:58:48.001078900+05:30[Asia/Calcutta]")
public class MemoryDTO {

  private String id;

  private String title;

  private String description;

  private Mood mood;

  @Valid
  private List<String> tags;

  @Valid
  private List<@Valid Image> images;

  public MemoryDTO() {
    super();
  }

  /**
   * Constructor with only required parameters
   */
  public MemoryDTO(String id, String title, String description, Mood mood) {
    this.id = id;
    this.title = title;
    this.description = description;
    this.mood = mood;
  }

  public MemoryDTO id(String id) {
    this.id = id;
    return this;
  }

  /**
   * Unique id for the memory
   * @return id
  */
  @NotNull 
  @Schema(name = "id", example = "123sdgt1315", description = "Unique id for the memory", requiredMode = Schema.RequiredMode.REQUIRED)
  @JsonProperty("id")
  public String getId() {
    return id;
  }

  public void setId(String id) {
    this.id = id;
  }

  public MemoryDTO title(String title) {
    this.title = title;
    return this;
  }

  /**
   * Short title for the memory
   * @return title
  */
  @NotNull 
  @Schema(name = "title", example = "First Day at New Job", description = "Short title for the memory", requiredMode = Schema.RequiredMode.REQUIRED)
  @JsonProperty("title")
  public String getTitle() {
    return title;
  }

  public void setTitle(String title) {
    this.title = title;
  }

  public MemoryDTO description(String description) {
    this.description = description;
    return this;
  }

  /**
   * Detailed description of memory
   * @return description
  */
  @NotNull 
  @Schema(name = "description", example = "Started my new role as a backend developer and met the team.", description = "Detailed description of memory", requiredMode = Schema.RequiredMode.REQUIRED)
  @JsonProperty("description")
  public String getDescription() {
    return description;
  }

  public void setDescription(String description) {
    this.description = description;
  }

  public MemoryDTO mood(Mood mood) {
    this.mood = mood;
    return this;
  }

  /**
   * Get mood
   * @return mood
  */
  @NotNull @Valid 
  @Schema(name = "mood", requiredMode = Schema.RequiredMode.REQUIRED)
  @JsonProperty("mood")
  public Mood getMood() {
    return mood;
  }

  public void setMood(Mood mood) {
    this.mood = mood;
  }

  public MemoryDTO tags(List<String> tags) {
    this.tags = tags;
    return this;
  }

  public MemoryDTO addTagsItem(String tagsItem) {
    if (this.tags == null) {
      this.tags = new ArrayList<>();
    }
    this.tags.add(tagsItem);
    return this;
  }

  /**
   * Optional labels used for categorizing or searching memories.
   * @return tags
  */
  
  @Schema(name = "tags", description = "Optional labels used for categorizing or searching memories.", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
  @JsonProperty("tags")
  public List<String> getTags() {
    return tags;
  }

  public void setTags(List<String> tags) {
    this.tags = tags;
  }

  public MemoryDTO images(List<@Valid Image> images) {
    this.images = images;
    return this;
  }

  public MemoryDTO addImagesItem(Image imagesItem) {
    if (this.images == null) {
      this.images = new ArrayList<>();
    }
    this.images.add(imagesItem);
    return this;
  }

  /**
   * 0 or more images associated with memory
   * @return images
  */
  @Valid 
  @Schema(name = "images", description = "0 or more images associated with memory", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
  @JsonProperty("images")
  public List<@Valid Image> getImages() {
    return images;
  }

  public void setImages(List<@Valid Image> images) {
    this.images = images;
  }

  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    MemoryDTO memoryDTO = (MemoryDTO) o;
    return Objects.equals(this.id, memoryDTO.id) &&
        Objects.equals(this.title, memoryDTO.title) &&
        Objects.equals(this.description, memoryDTO.description) &&
        Objects.equals(this.mood, memoryDTO.mood) &&
        Objects.equals(this.tags, memoryDTO.tags) &&
        Objects.equals(this.images, memoryDTO.images);
  }

  @Override
  public int hashCode() {
    return Objects.hash(id, title, description, mood, tags, images);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class MemoryDTO {\n");
    sb.append("    id: ").append(toIndentedString(id)).append("\n");
    sb.append("    title: ").append(toIndentedString(title)).append("\n");
    sb.append("    description: ").append(toIndentedString(description)).append("\n");
    sb.append("    mood: ").append(toIndentedString(mood)).append("\n");
    sb.append("    tags: ").append(toIndentedString(tags)).append("\n");
    sb.append("    images: ").append(toIndentedString(images)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}